refactor(tms): 将港口代码类型从 Integer 改为 Long
- 修改了 TmsFirstMilePageReqVO、TmsVesselTrackingDO、TmsVesselTrackingPageReqVO、TmsVesselTrackingRespVO 和 TmsVesselTrackingSaveReqVO 中的港口代码字段类型
- 移除了字段描述中的"（字典）"字样，简化为港口名称- 这个改动统一了港口代码的字段类型，提高了代码的一致性和可维护性

+++ yudao-module-tms/yudao-module-tms-biz/src/main/java/cn/iocoder/yudao/module/tms/dal/dataobject/vessel/tracking/TmsVesselTrackingDO.java
@@ -75,17 +75,17 @@ public class TmsVesselTrackingDO extends TenantBaseDO {
     @Version
     private Integer revision;
     /**
-     * 中转港（字典）
+     * 中转港
      */
-    private Integer transitPort;
+    private Long transitPort;
     /**
-     * 目的港（字典）
+     * 目的港
      */
-    private Integer toPort;
+    private Long toPort;
     /**
-     * 装运港（字典）
+     * 装运港
      */
-    private Integer fromPort;
+    private Long fromPort;
     /**
      * 船公司（供应商id）
      */
